'Brute force' changed to use ProofState.
Well, THAT was a lot harder than it should have been.

But, I did add a lot of types.  Yay, Python's type hints!  However, it turns out
mypy is a little "immature," so I had to change a named tuple to a regular
class.  But at least it can handle @property, so I could keep the immutability.

Next up: get it actually passing the type checker, also fix the non-brute-force
implementation which I'm sure I broke.